Add new move commands bound to C-; and C-:
Those just work like C-. or C-, i.e. move down or up a
paragraph. Except if a programming language major mode is active,
at which point it moves down or up the class/function definition. At
the moment, only python-mode is supported.

The move commands are included in the hydra move macro.